Model, do przetestowania

+++ app/src/main/java/com/siema/morse/model/Dash.java
@@ -8,7 +8,7 @@ package com.siema.morse.model;
  * 
  */
 public class Dash extends SoundChar {
-	public Dash(double time){
+	public Dash(long time){
 		super(3*time);	
 	}
